+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [1.4.0](https://github.com/ni/systemlink-grafana-plugins/compare/v1.3.0...v1.4.0) (2023-07-20)
+
+
+### Features
+
+* Add plotly panel ([cdd14a0](https://github.com/ni/systemlink-grafana-plugins/commit/cdd14a0fd12289599c069bdf5653299be6a27d30))
+
+# [1.3.0](https://github.com/ni/systemlink-grafana-plugins/compare/v1.2.0...v1.3.0) (2023-07-19)
+
+
+### Features
+
+* Add notebook data source ([905955a](https://github.com/ni/systemlink-grafana-plugins/commit/905955ab925903b9f0101cc857425af24e1c5945))
+
+# [1.2.0](https://github.com/ni/systemlink-grafana-plugins/compare/v1.1.0...v1.2.0) (2023-07-19)
+
+
+### Features
+
+* Systems data source template ([#3](https://github.com/ni/systemlink-grafana-plugins/issues/3)) ([6e3c79e](https://github.com/ni/systemlink-grafana-plugins/commit/6e3c79e569669a0dc69de50bd7f2173730d0f759))
+
+# [1.1.0](https://github.com/ni/systemlink-grafana-plugins/compare/v1.0.0...v1.1.0) (2023-07-19)
+
+
+### Features
+
+* Add tag data source prototype ([a2a04e6](https://github.com/ni/systemlink-grafana-plugins/commit/a2a04e6233a6c641b03eb0fccb4f97e48fdcfe2c))
+
 # 1.0.0 (2023-07-13)